# docs/swagger.yaml
basePath: /
definitions:
  auth.Role:
    properties:
      created:
        type: string
      created_by:
        type: string
      id:
        type: integer
      name:
        type: string
      priority:
        type: integer
      status:
        type: string
      updated:
        type: string
      updated_by:
        type: string
    type: object
  auth.RoleNew:
    properties:
      name:
        maxLength: 64
        minLength: 1
        type: string
      priority:
        minimum: 1
        type: integer
      status:
        enum:
        - 1
        - 2
        type: integer
    required:
    - name
    - priority
    - status
    type: object
  auth.SigninRequest:
    properties:
      auth_type:
        enum:
        - 1
        - 2
        type: integer
      credential:
        minLength: 6
        type: string
      identifier:
        type: string
    required:
    - auth_type
    - identifier
    type: object
  auth.SigninResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/auth.User'
    type: object
  auth.SignupRequest:
    properties:
      credential:
        minLength: 6
        type: string
      identifier:
        type: string
      organization_id:
        minimum: 1
        type: integer
    required:
    - credential
    - identifier
    - organization_id
    type: object
  auth.User:
    properties:
      address:
        type: string
      birthday:
        type: string
      created:
        type: string
      created_by:
        type: string
      credential:
        type: string
      email:
        type: string
      gender:
        type: string
      id:
        type: integer
      identifier:
        type: string
      name:
        type: string
      organization_id:
        type: integer
      phone:
        type: string
      position_id:
        type: integer
      role_id:
        type: integer
      status:
        type: integer
      type:
        type: integer
      updated:
        type: string
      updated_by:
        type: string
    type: object
  auth.UserUpdate:
    properties:
      address:
        minLength: 1
        type: string
      birthday:
        minLength: 1
        type: string
      email:
        type: string
      gender:
        minLength: 1
        type: string
      name:
        minLength: 2
        type: string
      phone:
        minLength: 1
        type: string
      position_id:
        minimum: 1
        type: integer
      role_id:
        minimum: 1
        type: integer
      status:
        minimum: 1
        type: integer
    type: object
...
  response.ErrorRes:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  response.ListRes:
    properties:
      count:
        type: integer
      data: {}
      page_id:
        type: integer
      page_size:
        type: integer
    type: object
  response.SuccessRes:
    properties:
      data: {}
    type: object
host: 0.0.0.0:8080
info:
  contact:
    email: lxx0103@yahoo.com
    name: Lewis
  description: API for zoho-center.
  title: zoho-center API
  version: "1.0"
paths:
...
  /roles:
    get:
      consumes:
      - application/json
      operationId: "18"
      parameters:
      - description: 页码
        in: query
        name: page_id
        required: true
        type: integer
      - description: 每页行数（5/10/15/20）
        in: query
        name: page_size
        required: true
        type: integer
      - description: 角色名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ListRes'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/auth.Role'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorRes'
      summary: 角色列表
      tags:
      - 角色管理
    post:
      consumes:
      - application/json
      operationId: "19"
      parameters:
      - description: 角色信息
        in: body
        name: role_info
        required: true
        schema:
          $ref: '#/definitions/auth.RoleNew'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessRes'
            - properties:
                data:
                  $ref: '#/definitions/auth.Role'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorRes'
      summary: 新建角色
      tags:
      - 角色管理
  /roles/:id:
    get:
      consumes:
      - application/json
      operationId: "20"
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessRes'
            - properties:
                data:
                  $ref: '#/definitions/auth.Role'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorRes'
      summary: 根据ID获取角色
      tags:
      - 角色管理
    put:
      consumes:
      - application/json
      operationId: "21"
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 角色信息
        in: body
        name: role_info
        required: true
        schema:
          $ref: '#/definitions/auth.RoleNew'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessRes'
            - properties:
                data:
                  $ref: '#/definitions/auth.Role'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorRes'
      summary: 根据ID更新角色
      tags:
      - 角色管理
  /signin:
    post:
      consumes:
      - application/json
      operationId: "17"
      parameters:
      - description: 登录类型
        in: body
        name: signin_info
        required: true
        schema:
          $ref: '#/definitions/auth.SigninRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessRes'
            - properties:
                data:
                  $ref: '#/definitions/auth.SigninResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorRes'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorRes'
      summary: 登录
      tags:
      - 用户权限
  /signup:
    post:
      consumes:
      - application/json
      operationId: "22"
      parameters:
      - description: 登录类型
        in: body
        name: signup_info
        required: true
        schema:
          $ref: '#/definitions/auth.SignupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessRes'
            - properties:
                data:
                  type: integer
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorRes'
      summary: 用户注册
      tags:
      - 用户权限
  /users/:id:
    put:
      consumes:
      - application/json
      operationId: "23"
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户信息
        in: body
        name: menu_info
        required: true
        schema:
          $ref: '#/definitions/auth.UserUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessRes'
            - properties:
                data:
                  $ref: '#/definitions/auth.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorRes'
      summary: 根据ID更新用户
      tags:
      - 用户管理
swagger: "2.0"